before changing calendar

+++ Airbnb/app/src/main/java/com/rp2/star/airbnb/src/main/search/searching/models/ResultLodgeDetailLodgeSpace.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class ResultLodgeDetailLodgeSpace(
     @SerializedName("id") val id: Int,
     @SerializedName("category") val category: String,
-    @SerializedName("amount") val amount: Int,
+    @SerializedName("amount") val amount: Float,
     @SerializedName("name") val name: String
 )
